2014 THE IMPACT OF USING TABLETS AS A PEDAGOGICAL TOOL FOR FIFTH  GRADE STUDENTS
 Abstract 
 
Technology plays an important role in current education. Many institutions have 
implemented the use of tablets as a pedagogical tool, replacing the traditional method of 
learning. This year, fifth grade students with a concentration in English at a private school in 
Cartagena   were provided tablets as a pedagogical tool, replacing Science, Math, Spanish, Social 
Studies and Religion printed textbooks. The content of these digital modules are not different 
from those printed modules used last year, except the images that were adapted with colorful 
drawings, and the activities were presented in an interactive way. Thus capitalizing on the 
technology that students enjoy and spend a great deal of time using. In respect to English class, 
tablets are used only as a way to support the class, like other common and important aids such as 
DVDs, video beam, CD players, etc. After four months of being using tablets in the classroom 
,students have had tablets in the classroom and the results are not the best. Students use the 
tablets to play non-academic games and they prefer engaging with the tablet as company instead 
of an activity with their classmates or teachers. With the aim of knowing the impact of using 
tablets in the classroom, this research approach students seek to understand their expectations 
and habits of using the tablets, then a lesson plan was structured to intervene the activities of the 
kind of English, then buscandose assess and capture the perceptions of students regarding this 
experience again .It covers 15 class hours by using tablets which will contain attractive app’s 
selected by teachers for students to learn teachings during the second term while having fun with 
interactive games.  
Key words: apps, tablets, interactive games, digital books, motivation.