2014 THE USE OF ROLE-PLAYS AND REAL LIFE-LIKE INTERACTION IN CLASS TO HELP STUDENTS DEVELOP EFFECTIVE COMMUNICATIVE SKILLS.
Abstract  
This pedagogical project aims at helping students improve their speaking skills 
through the use of real life-like interaction and role plays in L2, these have a socio 
cultural perspective in which learning is socially constructed by means of face-to-face 
contact among students and the context around them (Appel & Lantolf, 1994) 
Both practices are supported in Vygotsky’s Socio cultural theory, which states 
that the language is to be learned through interaction between peers one of them more 
capable than the others, since language acquisition is enhanced in environments where 
scaffolding can take place. 
This research project was applied at an institute in Cartagena where people go to 
learn English as a foreign language. Throughout the presentation of this project you will 
find all the information regarding the obstacles found along the way, as well as the 
solutions given to those problems. As you read it you will also notice how the 
Vygotskyan theory of scaffolding and ZPD (Zone of proximal development) raises 
awareness of the importance of interaction in class. Especially when comes to helping 
students go beyond their actual scope of knowledge and competence in L2.  The results 
of the work are visible in the final outcomes of those students that were part of this 
project from the beginning.  
Key words: Scaffolding, ZPD, peer interaction.