2018PROMOTING LEARNERS TO IMPROVE THEIR SPEAKING SKILL THROUGH GAMES IN A PUBLIC SCHOOL.
ABSTRACT
This study presents the strategies that foster oral production of seventh students in a public school of Cartagena. It is the result of an action research, which responds to the descriptive and interpretative method. The aim of the current work is to investigate the impact that game activities produce on the development of learners´ speaking skills. Three instruments were used for the collection of information, namely: observations, digital survey and an interview in order to find the speaking skills students develop with the implementation of games. The data gathered showed that the students want to improve their speaking skill. Hence the necessity to look for strategies to encourage students to enhance this communicative skill.
A pedagogical intervention was implemented. This consisted of four lessons aimed to foster learners‟ use of games to develop their speaking skill. The results of the data analysis reveal that the participant group‟s confidence has increased, as well as their speaking level.
Keywords: Speaking, games, advantages of games, types of games.