2015ROLE PLAY AND DISCUSSION QUESTIONS AS STRATEGIES TO IMPROVE THE SPEAKING SKILL IN TOURISM AND LANGUAGES STUDENTS AT INSTITUCIÓN TECNOLÓGICA COLEGIO MAYOR DE BOLIVAR.
Abstract
This small scale research was developed at Institución Tecnológica Colegio Mayor de Bolivar, with third semester students from the tourism and languages program. The main objective of this project is to improve the speaking skill of the students through the implementation of role plays and discussion questions.
To develop this research we started with the identification of the problem; then the collection, organization and interpretation of data in order to plan actions that help us achieve the objective, and finally we wrote some conclusions according to the results.
We found out that both role plays and discussion questions can serve as vehicles to enhance the student’s speaking skill, especially if, when implemented, those strategies are applied in accordance to the findings of this paper.
Keywords: Speaking, role plays, discussion questions, improvement, class participation.