2017 A SIOP COURSE FOR IN SERVICE TEACHERS OF THE BACHELOR’S LANGUAGE TEACHING PROGRAM AT FUNDACIÓN UNIVERSITARIA COLOMBO INTERNACIONAL, UNICOLOMBO
ABSTRACT
The main purpose of this study was to assist Education college teachers from the Unicolombo
Teaching Program in developing competencies in order to carry out CBI lessons using the SIOP
model as an alternative and to be capable of integrating language and content simultaneously
through the design and implementation of a SIOP course. Questions formulated to achieve the
purpose of the study focused on. (1) Diagnosing what SIOP features teachers apply in their
classes. (2) Implementing a SIOP course regarding the professor’s needs. (3) Describing the
improvement of the Licenciatura faculty’s SIOP pedagogical knowledge. Eleven teachers
constituted the sample for this study and three instruments designed by the researchers were used
for data collection, which were surveys, worksheets, and lesson plans. Results indicated that
teachers made noticeable modifications in their initial lesson plans in terms of learning strategies,
scaffolding techniques, assessment, and the integration of language objectives in accordance
with content objectives. The effect of Sheltered Instruction Observation Protocol model on
teachers’ teaching competencies was somehow significant. This course gave teachers the
opportunity to reflect on their pedagogical practice in order to make decisions and be creative.
Key words: Sheltered Instruction Observation Protocol model, teaching competencies, CBI
lessons, content and language integration, teacher training.
2