2014 The Effect of a Dual Intervention on University Students’ Academic Writing Skills: A Case
Abstract
Fifth semester students of the B.A in education with a focus in English at a private university in
Cartagena take content-based courses, among which is Second Language Acquisition II (SLA
II). One of the requirements in this course is to write a reflection paper. However, the experience
of one of the teacher-researchers who participated in this study reveals that at the beginning of
these courses, most students seem to lack the necessary academic reading and writing skills to
complete the assignments properly. This action research describes the effect of a dual
intervention that includes a 15-hour pre-semester course in academic writing and an academic
writing course in Moodle (a free e-learning platform) on the academic writing skills of 16
students in a SLA II class. Two surveys were administered to them: one right after they attended
the classroom-based course. The other one after they submitted their final draft of their reflection
papers. Both surveys included issues related to the effectiveness of the intervention. Then, their
final versions of the papers were evaluated and compared to those written by another group.
Findings suggest that the intervention was helpful to a certain extent. Further research is needed
to explore alternative ways of tackling this issue.
Key words: content-based instruction, academic writing, writing skills, reflection papers.