2016 CBI SYLLABUS ON MUSICAL NOTATION FOR INSTITUCION UNIVERSITARIA BELLAS ARTES Y CIENCIAS DE BOLÍVAR
Abstract
This research Project was developed with the aim of creating a practical and innovate micro curriculum
for the English classes at Institución Universitaria Bellas Artes y Ciencias de Bolívar (UNIBAC) At the
University, there was not a previous curriculum which taught content and language at the same time, and
their policies required one whose objective pointed towards internationalization.
To carry out this work an interviewing the students and the directors of the music program to get relevant
information about their needs and expectations on the English language, we went through different
theories of learning to collect valuable facts about learning languages involving content and language and
we studied models alike such as CBI, Theme based and SIOP.
The music program at UNIBAC has been making efforts in order to achieve a good musical position at
the international level through tours, concerts, international workshops, among other activities, in these
activities English plays an important role for communication among the participants, however, It lacks of
a micro curriculum in the education part, this micro curriculum does not include the contents and the
language at the same time. To accomplish this, the authors decided to propose one in which the necessary
musical contents are present to learn and practice the lexical part, grammar, and content; this will allow
the institution the learning of the language.
To do this, books and articles about CBI, theme-based instruction, SIOP were chosen; along with theory
of harmony and music contents, descriptors of the level A1 form the Common European framework,
language competence, and the suggested curriculum of the national ministry of education. This
curriculum should be adapted to the objectives of the university. The results show the use of the language
and the musical contents that are appropriate to be developed in the level A1 and the use of SIOP as
teacher`s guide to prepare the lessons. This work aims at contributing to the education in the English
language for music students from the music theory level A1 accompanied by teachers who are not
specialists in the topic.
Key words: CBI, Theme based learning, SIOP, A1 Level, Music theory.