2017 STRENGTHENING WRITING SKILLS BY USING COHESIVE DEVICES IN COURSE 2 STUDENTS AT CENTRO CULTURAL COLOMBO AMERICANO THROUGH WRITTEN TEXTS
ABSTRACT
This study examines the difficulty students present when communicating their daily activities throughout written texts in a clear and proper way. A common problem found among students from the English program of Centro Cultural Colombo Americano (CCCA) is the lack of writing skills regarding sentences and paragraph construction. Hence, it is imperative to look at this situation closely in order to find a solution and help students to improve their communicative skills.
The investigation was carried out at CCCA, eight students in Course 2 were selected to be part of this Action Research Project. Through a survey applied to students, they expressed that they would like to improve their writing skill, but feel anxiety due to the deficiency of appropriate vocabulary to convey their ideas during the text construction process.
In order to overcome this situation, it was determined to strengthen students‟ writing skill through the use of cohesive devices and mechanics, as well as the scaffolding technique, due to its suitability for the purpose.
All the way through this research, it was possible to witness the efficacy of scaffolding as a teaching strategy to boost L2 students‟ acquisition, not only on early stages but in any stage of the learning process.
KEYWORDS
Writing skills, cohesive devices, Scaffolding, Action Research