2017 Enhancing Literacy Ability in English beginner Students of Instituto Bolivariano Esdiseños through the Use of Reading Strategies
Abstract
This study examines the effects of reading strategies instruction in EFL beginners with the purpose of enhancing literacy. This study was carried out in a vocational purpose institution located in Cartagena de Indias. It attempts to undertake the succeeding question How to enhance in English beginner students of instituto Bolivariano Esdiseños the literacy ability through the use of reading strategies?
To develop this research, we began with the identification of the problem; then, the data collection through the use of two instruments: observations, and a survey; the organization and interpretation of data in order to implement the action plan addressed through interventions, which helped to achieve the main research objective. Finally, we wrote some conclusions according to the results.
Our findings suggest that through the implementation of explicit reading strategy instruction students are able to enhance their literacy ability that enable them to develop lifelong learning, critical thinking skills, a sense of autonomy, and abilities to face the challenges of the reading competence.
Key words: Reading skill, Literacy, Reading strategies.