2015 IMPLEMENTATION OF COMMUNICATIVE ACTIVITIES TO IMPROVE THE SPEAKING SKILLS ON LIBRARIANS AT A BILINGUAL UNIVERSITY IN CARTAGENA
ABSTRACT
This case-study research intends to achieve oral communication skills through conducting activities for librarians to improve their speaking skills in a private bilingual university in Cartagena, the techniques implemented to collect data of the activities were surveys, video and the journal.
After piloting the lessons, the problem showed that librarians had difficulties to attend people who asked for the library service in English, to solve it, we formulated how the implementation of communicative activities can improve the speaking skills on librarians?
Since the objective is to have librarians communicate, then they should be motivated to interact by improving their speaking skills, through the implementation of ESP lessons for librarians. We emphasized on communication and interaction, we carried out a learner centered approach that suit with the needs of that specific context. Xhaferi and Xhaferi (2011) said that “It is widely believed that language skills reading, writing, listening and speaking are very important in teaching ESP courses but it is often depends upon the leaner‟s needs and interests” (p.435).
ESP lessons were designed to develop strategies to offer the service and guide library users in English.
Key words: ESP speaking activities learner centered approach speaking skills