2014 DEVELOPING A READING SKILLS UNIT FOR B1 LEVEL STUDENTS AT THE CENTRO COLOMBO AMERICANO CARTAGENA
INTRODUCTION
English has become an important universal language, it is a communication tool which opens many doors. Many people are learning English in order to become more competent professionals in this world. In addition to that, universities are also offering English classes to fulfill the high demand of people who are required to learn this language. The Centro Colombo Americano in Cartagena is a great influence in Colombia because it has been contributing to this purpose of teaching English as a second language for almost 50 years.
The Centro Colombo Americano in Cartagena is a non-formal educational English institution with a non-profit making objective, located in the north of Cartagena. This institution offers the opportunity of learning English for the general public. Approximately, five years ago, a new branch was opened in Los Cuatro Vientos neighborhood in order to contribute to improving the quality of life of Cartagena`s society.
The Colombo in Cuatro Vientos is located half a block from Pedro de Heredia Avenue. This is a middle class neighborhood which is sorrounded by some low class areas.
In this neighborhood, there is a well-known furniture store called “Muebles Jamar” and we can notice the different discourses in which the students are involved at this site in Los Cuatro Vientos, such as a great deal of restaurants, car-part stores, grocery stores, a bakery ,just across from the Colombo; as well as some schools, a health center and other businesses. This area has heavy traffic almost all the time which might be a little dangerous for the students who frequently have to cross some streets and the main avenue to go home. Some of the students at the Colombo prefer to stay around Los Cuatro Vientos to have lunch because they do not have time or money to go back home, as a result, most of them take “mototaxi” service to returm home since it is fast and a little cheap for them.
The average Colombo students belong to the middle class society with some lacks and needs in common and a very popular culture. Their ages vary from fifteen to forty and their economical situation is mostly acceptable. Regarding the learners; the ages vary. There are young adults and adults. The aim of the English program is the development of the English language for general purposes. The course starts in a A1 level and intends to take the students to minimum a B2 level on the completion of the program according to The Common European Framework, Council of Europe (2001) .
This project is carried out in Cuatro Vientos in Cartagena, which is another Colombo branch surrounded by some poor neighborhoods from which many students come and most of them show reading difficulties. There are many signs that students have to read almost everyday and some of them are progressively becoming bilingual since we live in a tourinst city, therefore, these bilingual literacy practices should be considered important by Colombo English students.
The main reason for choosing this institution for my project is because reading is an important skill in teaching and learning a second language and it has been observed that this skill is often neglected by some English teachers here.
Students appear to have limited or reading exposure in English classes. Therefore, most of them seem to have trouble understanding and discussing topics when they read different texts in English. Besides, it appears that they need to be encouraged to read more in order to increase their abilities.
The needs analysis carried out at this Colombo branch shows the immediate improvement of this skill. In a survey, A lot of the students claim that they do not feel prepared to give their opinion after reading certain texts in English. In addition, All of them agree that reading is an important skill because they can learn and understand more, they say it is as necessary in english as it is in their mother tongue. In the survey, the students also state that vocabulary is the most complex section when reading an English text. Almost all the students are in favor of implementing this reading unit. As a result, developing this project could help students have a good command of their reading skills and become more competent English speakers at the Centro Colombo Americano.
Our research question is “What effects can developing a reading skills unit have on B1 students to improve reading comprehension at the Centro Colombo Americano in Cartagena?” This project will be done by designing this unit which will be offered to those students who need to develop communicative reading skills though contextualized topics of students` interests. According to the survey, students assert that they enjoy reading different types of texts, especially stories.
The main objective of this is to identify and describe the effects of this reading skills unit offered to thirty 17 - 45 year old people from two B1 level classes with which we work at the Centro Colombo Americano in Cartagena (Bolivar).