2014 ANALYZING DIFFERENT STRATEGIES WHICH HELP FE Y ALEGRIA SCHOOL’S TENTH GRADERS EXPRESS ORALLY IN ENGLISH
ABSTRACT
This investigation is focused in the Different Strategies Which Help Fe Y Alegría School‟s Tenth Graders to Express Orally in English. Besides of identifying and analyzing different strategies which help to improve students‟ speaking skills based on an interactional perspective by developing this skill by carrying out different kind of interview, role plays, and small talks which let students experience the reality of the topics taught.
In this research the unstructured observation was applied as method to gather information, because it is selective in focus and it can be easily adapted to the individual Classroom situation. Unsystematic or unstructured observation is one that takes place in an open, unstructured, unsystematic and without the use of special tools to collect the data. It is usual in the qualitative research and is used by sociologists, anthropologists and ethnographers in their field research, which seeks by a more direct way, the knowing in depth way some realities of individuals, groups and communities.
They were taken six activities and transcribed to a format (unstructured) the dates: 1. Context identification, 2. objective, 3. description of the situation (numbers of observed students), teacher description, time, conclusions etc), 4. Columns with dates as: language, turn, actor, events, and found categories. 2(First, it is important to say if the interaction was taken place and besides if the interaction was unidirectional or bidirectional).
The observed object was a group of 30 teenagers (20 males and 10 females) between the ages of 14 and 16 with a basic level of English at an official school of Cartagena, studying 4 hours per week.
The clue words that guided this research project were:
 Interaction
 Learning Strategies
 Oral skill
 Activities