2014 SPECIAL L2 LEARNING PROCESSES IN STUDENTS WITH AND WITHOUT NEEDS
ABSTRACT
Montessori is a private bilingual school located in Manga neighborhood which is a high class community surrounded by different kinds of places such as banks, parks, restaurants, schools and is right next to Cartagena bay, the school provides a good quality of education based on Montessori’s methodology.
The increase of students with learning difficulties has caught the attention of the school communities, reason why this research project is focused on L2 learning process in students with and without special needs and also on the kind of strategies that can be used with them.
As a strategy, we proposed the following lesson plan which aim to foster L2 learning through inclusive activities promoting interactive learning on students with and without special needs. (See below)
Week
Date
Description of activity/Workshop
1
April 21st to April 25th
Workshop
1. Match the image of the place of the city with the word written.
2. Think about your neighborhood and draw the places that you remember
3. Fill the blanks with the vocabulary learned
3
As a result of the application of this lesson plan, students have shown an active attitude to resolve it, including a student with special needs that was in the group, the students without special needs were supported by the student in mention demonstrating that there is no exclusion in a social perspective. In addition, the student with a specific diagnostic was encouraging the regular students to develop this activity
Bilingualism
Special needs
Inclusion education