2016 Running head: SPECIFIC ORAL SKILLS FOR 7th GRADERS AT GAC
Abstract
How to Become an Excellent Presenter
The center and heart of this research project was to identify what specific oral skills were needed in order to be taught, reinforced and improved in a group of seventh graders students at Gimnasio Altair de Cartagena. Specifically the project was set in the abilities required in their oral presentations in front of a class. To address this problem, a need analysis was carried out by applying different research techniques and tools such as questionnaires, surveys, focus groups and class observation, which emitted certain research information that lead us to approach our problematic. As a result of this process and after the application of the academic program designed after our research findings set to improve their oral skills, our students showed a high level of improvement as English speakers, an increment in their self-confidence at the moment of standing in front of their class was also noticed and showed themselves more secure when presenting in front of their peers using the techniques they were taught.
Key words: oral skills, abilities to speak, oral improvement, oral presentations