2014 EL JUEGO COMO RECURSO LÚDICO EN EL DESARROLLO DE LA ESCRITURA EN IDIOMA INGLÉS DE TERCER GRADO DE LA INSTITUCIÓN EDUCATIVA TÉCNICA LA BOQUILLA SEDE MADRE BERNARDA
ABSTRACT
In this research project a proposal is presented which includes the game as a ludic resource aim to the development of the written production skill in the English language learning process of third grade students of Técnica la Boquilla sede Madre Bernarda School. To achieve this objective initial students’ written production weaknesses were identified through a diagnostic test. In order to design the test that was applied to students, some basic standards of foreign language competences which correspond to the writing skill such as recognizing elements in a picture, copying in English and writing personal information in simple formats. The result obtained after the implementation of pedagogical proposal evidenced that students could overcome their initial writing production weaknesses in English.
Key words: game, meaningful learning, ludic, English, written production.