2014 EL USO DEL ESPAÑOL EN EL PROCESO DE ENSEÑANZA - APRENDIZAJE DEL INGLÉS EN TERCER GRADO DE LA INSTITUCIÓN EDUCATIVA FOCO ROJO DE CARTAGENA
Abstract
Cartagena Being a tourist town and wanting to strengthen tourism and improve the quality of life, the National Bilingual Program was created for the teaching of English in order to understand and manipulate the target language in different contexts: social, cultural and labor.
Our public education system is in crisis in teacher training in teaching English in elementary school and notes that these are not professional in this specific area.
In the interview the newspaper El Tiempo (2009), Ms. Rosa Cely - Academic Advisor to the National Plan of Bilingualism, said that the English classes being run mostly in Spanish and this was demonstrated by the low results assessments that measure the quality of learning English because the Spanish used inappropriately becoming a negative tool because it relieves the student of the correct practice of English because the need to communicate in this language is created.
This paper analyzed the inappropriate use of Spanish in the classroom Grade 3rd of School Focus Red in Cartagena through a case study from 2011 to 2013 where the suitability of the use of L1 was studied in the learning of English in the students of this institution.
The methodology of this research was supported by the interpretive qualitative paradigm, where it was understood and interpreted the times when the Spanish used in English classes, and the opinions, practices or processes related to their use from the point of view of teachers and students. To this end the techniques used in this research were observation, interview and survey and the results of our interventions in the English classes, was that it showed a little interest in students and teachers use the language foreign in their classes for this to be more meaningful and see it as an opportunity to better prepare for their pedagogical.
Keywords: Mother language or L1, Foreign language, Second language, teaching – learning process.