2013 LÚDICA COMO ESTRATEGIA PEDAGÓGICA PARA PROMOVER LA AFECTIVIDAD EN LA ADQUISICIÓN DEL IDIOMA INGLÉS EN LA INSTITUCIÓN EDUCATIVA FOCO ROJO
Abstract
The goal of the following paper was to promote affectivity, among the eighth grade students, in the English language; the pedagogic strategy used in order to facilitate the acquisition of a second language was the playful.
This research was carried out under the theoretical approaches of Nájera and Jimenez. The first author defends the idea of affectivity and its connection with pedagogical issues, whereas the second author advocates the use of playful. The collection of data was carried out by the field diary. It was evident that the students showed some reluctance to English learning since the grammatical and translation activities plus the homework monotony turned out to be incomprehensible to them. The improvement plan showed that by making students feel important, and by transmitting them joy and comfort, the language acquisition process increases since the students feel more confident and learn, for instance, that we can learn through our own mistakes.
Keywords: feelings, emotions, playful and adquisition