2015 FORTALECIMIENTO  DE LA HABILIDAD ORAL DEL IDIOMA INGLÉS  A TRAVÉS DEL USO DE AUDIOS  Y  VIDEOS  CON  ESTUDIANTES DE DÉCIMO GRADO  DE LA MODALIDAD DE BILINGÜISMO DEL COLEGIO INEM
ABSTRACT 
This research had the objective to strengthen the English oral skills in the tenth grade students of 
Bilingualism class in the Institution Jose Manuel Rodriguez Torices through activities in which 
audios and videos were used. Starting with an initial diagnostic test in which the Bilingualism 
students found difficulties in the oral production, especially in aspects like interactive 
communication, pronunciation and vocabulary. In order to solve the problem encountered, this 
research was carried out in three stages or phases based on Participative Action-Research, called 
diagnostic phase, phase action plan development and implementation phase. At the conclusion of 
the last stage, a final diagnostic test was performed to compare the before and after of the 
investigation and determine whether the objectives of the research were made.  
Keywords: audios, videos, oral production skill, English, diagnostic test.