2017 A DIAGNOSIS OF ENGLISH LANGUAGE TEACHING MATERIALS FROM AN  INTERCULTURAL AND ECOLOGICAL PERSPECTIVE IN CARTAGENA DE  INDIAS, COLOMBIA
Abstract  
 
Intercultural and Ecological perspectives are the main approaches Colombian 
Ministry of Education proposed in order to enrich the pedagogical practices in English 
teachers and to achieve the primary goal of the Colombia Bilingual program which is to 
improve English language skills by 2025. In the view of that fact, it had been reached an 
exploration of the policies the Ministry of Education provides and a discussion of materials 
used in Cartagena’s school through a material evaluation checklist stated under 
Intercultural and Ecological features that illustrate the actual situation of the application of 
 low percentage of suitable materials during English classes. The research aims to supply 
information for teachers, concerning about how the process is going and what to do in favor 
of the bilingual country.  
 
Keywords: Interculturality, Ecological approach, materials.