2015 IMPLEMENTACIÓN DE LA EVALUACIÓN COMO ACTIVIDAD LÚDICOPEDAGÓGICA EN EL PROCESO DE APRENDIZAJE DEL IDIOMA INGLÉS EN  LOS  ESTUDIANTES DE 5 ° DE PRIMARIA DE LA I.E. PROMOCIÓN SOCIAL, SEDE LA CONSOLATA
Abstract 
 
     This research project focused on the implementation of a proposed recreational 
educational evaluation for Basic Elementary 5th grade I.E. Promoción Social, Sede La 
Consolata.  His goal was based on strengthening the learning of English. The research 
undertaken was descriptive qualitative-type intervention methodology action, which constantly 
seeks innovative learning strategies for motivating students. It addressed the problem of lack of 
assessment strategies by the teacher in charge of the group; in addition, to her lack of proficiency 
in English; which was evidenced through direct observation, structured field notes, surveys and 
interviews with teachers and students. Evaluation techniques developed in the class were 
observed and characterized as traditional, with the predominance of the written examination; 
this, perhaps, is a possible reason why students do not develop English skills. These findings led 
to the design of assessment strategies playful important elements such as reviving the everyday, 
looking from another perspective the student category - Professor, strengthen values, encourage 
rather than punish, use humor as a symbol of intelligence and recognize that what makes Playful 
a training and education for positive growth. 
The keywords:  Evaluation, Leisure, Activity, Teaching, and Learning