2012 MOTIVACIÓN EN LOS PROCESO DE ENSEÑANZA Y APRENDIZAJE DEL IDIOMA INGLÉS EN GRADO 8°1 DE LA INSTITUCIÓN EDUCATIVA  FOCO ROJO
ABSTRACT 
This research was developed at the Foco Rojo educational institution, located in an 
urban stratum 1 and 2 of Olaya Herrera district. This school has 341 students in 
the afternoon.  In each class, there are 35 students. The classrooms are poorly 
equipped. There are no windows or doors, which favors the lack of discipline. The 
class where the project is developed is the 8th grade with 38 students, 22 girls and 
16 boys, aged 15 to 17 years old. The researcher wants to know how students are 
motivated when they attend their English classes and their attitude towards it. For 
most students, there was an extrinsic motivation, they developed the tasks or 
activities just to meet a requirement of the institution to pass the subject so they 
can pass the school year and please their parents and not due to self-interest 
reasons such as acquiring new knowledge, learning about another culture or 
meeting foreigners. It was also important to identify the teaching strategies used to 
induce states of motivation to their students, Strategies that were necessarily 
implemented in order to create intrinsic motivation in students so that they lose the 
apathy of the English language and learn in a fun way through the different 
motivating dynamics. 
 
Keywords: Motivation, Education - Learning, English, Second Language 
Acquisition, Attitude, Strategies, Dynamics, Methodology.