2015 INTERACCIONES EN LA ENSEÑANZA-APRENDIZAJE DEL INGLÉS EN LA  INSTITUCIÓN EDUCATIVA DE TERNERA, CARTAGENA DE INDIAS, COLOMBIA
ABSTRACT  The interaction therefore it facilitates access to knowledge of the world of the other and of itself, essential in all social systems. Hence the need to recognize and study this in the educational field. This actions research has been developing since 2012 in public educational institutions of the city of Cartagena. This report presents partial results concerning the interactions between students and teachers of the Institución Educativa de Ternera, specifically sixth and eleventh grades. The observations and interventions basically revealed that the interaction is an essential factor in any process of acquisition of English as a foreign language directly affects the quality of education. Keywords: interactions, teaching and learning of English public school