2015 INTERCULTURALIDAD Y EDUCACIÓN Diseño de materiales para la enseñanza-aprendizaje  del idioma inglés desde enfoque intercultural: una revisión
ABSTRACT  The use of instructional materials allows to strengthen theoretical and practical aspects in the classroom, which has an impact on the development of strategies making possible the interaction providing benefits through the learning process; examining the use of instructional materials from an intercultural approach, suggests imminently the adoption of aspects that delimit and set differences on the very characteristics of this phenomenon, to achieve a meaningful advance in the communicative skills in the English teaching as foreign language and to produce changes in the traditional curricula and the critical awareness of the students about their cultural identity and appreciation of the differences in reference to other cultures from implementing the use of materials in education. The instructional materials need to have an initial purpose, should not appear out of nowhere. The teacher has to keep in mind the needs and special features of the context and the participants of the process without forgetting about the characteristics that define the culture of the target language; from quality to variety in designing materials, they require a detailed analysis in order to minimize mistakes.   
 
Keywords: English, interculturality, materials, instructional materials, culture in education.