2016 PROPUESTA DE UNA CARTILLA FUNDAMENTADA EN EL ENFOQUE CBI PARA LA CLASE DE  INGLÉS DEL GRADO SEXTO DE LA INSTITUCIÓN EDUCATIVA DE PONTEZUELA (FASE DE DISEÑO).
ABSTRACT 
 
This paper proposes the design of a booklet as material that support English teaching for sixth grade students of Pontezuela`s 
School. This booklet is based on theme model of content-based approach, CBI in English: Content Based Instruction. This approach 
is intended to teach English in interdisciplinary way in order to counteract the shortcomings that students have in sciences. 
 
 
 
Keywords: CBI; design; Science; approach; content, theme.