2015 CARTILLA  PARA EL MEJORAMIENTO DE LA COMPETENCIA COMUNICATIVA  DEL IDIOMA INGLES  DEL GRADO 4º DE  LA INSTITUCIÓN EDUCATIVA  OMAIRA  SÁNCHEZ GARZÓN
ABSTRACT 
    Every educator, when dealing with the teaching of English language class, must design 
or at least select the resources and materials he/she plans to use, in order to reach an 
interaction with his apprentices. Some teachers neither have teaching materials in their 
classrooms and nor show the importance of using them, knowing that the use of educational 
materials for the teaching of English in primary is very important, resulting in student 
motivation and ease of properly interaction with these educational resources. Therefore, 
research comes into play at Omaira Sánchez Garzón School in which the use of educational 
materials is provided through the design of an educational brochure to facilitate the 
development of the educational process which is supported by the National English 
Program in the country. In Cartagena, public educational institutions should implement the 
use of these essential tools for the development and enhancement of the teaching-learning 
process. 
Keywords: teaching materials, students, English language teaching.