2016 Improvement of High School Students’ Oral Skill Using Task Based Instruction
Abstract  
This paper   illustrates how a specific group of learners  that belongs to tenth and eleventh 
grade in high school improves their oral skills by learning through Task Based Instruction. Based 
on the action-participative research , firstly,  a speaking test is applied to students. The results of 
this evaluation shed lights about the oral English level of the students, Secondly, participant 
observations are carried out in TBI classes in order to know the performance of the students in 
this kind of classes.  It can be reported that in the speaking test, the learners presented several 
failures in the oral skill according to the level they were supposed to be. On the other hand,  the 
English classes based on TBI help students to engage and improve their oral competency. 
According to the present research, Task Based Instruction provides benefits to the learners since 
they are involved in the target language as they learn in a meaningful way. This makes students 
use the language and be able to participate in a conversation with a native person; expressing 
their ideas with  more confidence. 
 
 KEY WORDS: Task based Instruction, oral skill, young adults